+++ tests/test_util.py
@@ -1,14 +1,10 @@
 from __future__ import division, print_function, absolute_import
 
-from nnmnkwii.util import trim_zeros_frames, remove_zeros_frames
-from nnmnkwii.util import adjast_frame_length, apply_delta_windows
-from nnmnkwii.util import example_audio_file
+from nnmnkwii.util import apply_each2d_padded, apply_each2d_trim
 from nnmnkwii import functions as F
 from nnmnkwii.util.linalg import cholesky_inv, cholesky_inv_banded
 
-from scipy.io import wavfile
 import numpy as np
-import pyworld
 import scipy.linalg
 import bandmat as bm
 
@@ -33,57 +29,29 @@ def _get_windows_set():
     return windows_set
 
 
-def test_trim_remove_zeros_frames():
-    fs, x = wavfile.read(example_audio_file())
-    frame_period = 5
+def test_function_utils():
+    def dummmy_func2d(x):
+        return x + 1
 
-    x = x.astype(np.float64)
-    f0, timeaxis = pyworld.dio(x, fs, frame_period=frame_period)
-    spectrogram = pyworld.cheaptrick(x, f0, timeaxis, fs)
-    aperiodicity = pyworld.d4c(x, f0, timeaxis, fs)
+    T, D = 10, 24
+    np.random.seed(1234)
+    X = np.random.rand(2, T, D)
+    lengths = [60, 100]
 
-    for mat in [spectrogram, aperiodicity]:
-        trimmed = trim_zeros_frames(mat)
-        assert trimmed.shape[1] == mat.shape[1]
+    # Paddd case
+    Y = apply_each2d_padded(dummmy_func2d, X, lengths)
+    for i, l in enumerate(lengths):
+        assert np.allclose(X[i][:l] + 1, Y[i][:l])
+        assert np.all(Y[i][l:] == 0)
 
-    for mat in [spectrogram, aperiodicity]:
-        trimmed = remove_zeros_frames(mat)
-        assert trimmed.shape[1] == mat.shape[1]
+    # Trim
+    for i, l in enumerate(lengths):
+        X[i][l:] = 0
+    Y = apply_each2d_trim(dummmy_func2d, X)
+    for i, l in enumerate(lengths):
+        assert np.allclose(X[i][:l] + 1, Y[i][:l])
+        assert np.all(Y[i][l:] == 0)
 
-
-def test_adjast_frame_length():
-    D = 5
-    T1 = 10
-    T2 = 11
-
-    x = np.random.rand(T1, D)
-    y = np.random.rand(T2, D)
-    x_hat, y_hat = adjast_frame_length(x, y, pad=True)
-    assert x_hat.shape == y_hat.shape
-    assert x_hat.shape[0] == 11
-
-    x_hat, y_hat = adjast_frame_length(x, y, pad=False)
-    assert x_hat.shape == y_hat.shape
-    assert x_hat.shape[0] == 10
-
-    x_hat, y_hat = adjast_frame_length(x, y, pad=True,
-                                       ensure_even=True)
-    assert x_hat.shape == y_hat.shape
-    assert x_hat.shape[0] == 12
-
-    x_hat, y_hat = adjast_frame_length(x, y, pad=False,
-                                       ensure_even=True)
-    assert x_hat.shape == y_hat.shape
-    assert x_hat.shape[0] == 10
-
-
-def test_apply_delta_windows():
-    T = 5
-    static_dim = 2
-    x = np.random.rand(T, static_dim)
-    for windows in _get_windows_set():
-        y = apply_delta_windows(x, windows)
-        assert y.shape == (T, static_dim * len(windows))
 
 def _get_banded_test_mat(win_mats, T):
     sdw = max([win_mat.l + win_mat.u for win_mat in win_mats])
